fix: ensure generated conftest.py located in generated pytest folder
